Change habit status from days to percentage
The habit count was the number of consecutive days it was completed until the previous day. Counting this way, a single day missed resets the count, which I think discourages the user from maintaining their habits.

Calculating a percentage related to the number of times the habit was completed in the previous 20 days is better.