Allow the creation of plans inside the editor
This will create plans for new files inside the plans folder. The plan
name will be the file name without the ".md" extension, if it is present
and the name will not be empty if it is removed.

The plan order is kept for existing plans, the new ones are added in the
end of the list.