Increase abnormal time period warning limit
There was one time period longer that 10 hours in my usage of the tool.

The limit is now 20 hours.

This was created as a way to prevent time periods that are unusually
long so that the user can delete them if they are wrong.

There should be a way to warn the user without showing the warning
everytime there's a time period that passes the limit.